Translate x, y coordinates by delta x, delta y

Translate x, y coordinates by delta x, delta y

Usage

ST_Translate(tbl, geomName = "geom", dx, dy)

Arguments

tbl

name of a table in a duckdb database

geomName

name of the column containing the geometry value in the tbl. default = "geom".

dx

delta x

dy

delta y

Value

a duckdb table with translated geometries

Examples

con = DBI::dbConnect(duckdb::duckdb(), ":memory:")

value = system.file("data", "dummy_points.shp", package = "dbSpatial")

points <- ST_Read(conn = con, name = "points", value = value, overwrite = TRUE)
#> Error in .check_value(value): Invalid "value" input passed.

points |> 
  dplyr::mutate(geom_text = ST_AsText(geom))
#> Error in UseMethod("mutate"): no applicable method for 'mutate' applied to an object of class "function"

points_translated <- ST_Translate(tbl = points, dx = 100, dy = -20)
#> Error in .check_tbl(tbl = tbl): Please provide a duckdb table

points_translated |> 
  dplyr::mutate(geom_text = ST_AsText(geom))
#> Error in eval(expr, envir, enclos): object 'points_translated' not found