Create a duckdb table with spatial points

Create a duckdb table with spatial points

Usage

createDbSpatPoints(data, conn, tbl, overwrite = FALSE, ...)

Arguments

data

Data to be added to database. Can be a data.frame, tbl, file path, or terra spatVector/spatRaster. Experimental support for lists of files is supported through duckdbfs.

conn

Connection to duckdb database.

tbl

Name of table to add dbSpatPoints to.

overwrite

Overwrite existing table (default = FALSE)

...

Additional arguments to be passed

Value

A duckdb table with spatial points.

Details

For full list of supported file types see ST_Read() in duckdb.

Examples

# create in-memory duckdb db
duckdb_conn = DBI::dbConnect(duckdb::duckdb(), ":memory:")

# test data
test_data = data.frame(x = 1:10, y = 1:10, id = 1:10)

write.csv(test_data, "test_data.csv", row.names = FALSE)

# df, tbl
createDbSpatPoints(conn = duckdb_conn,
                   tbl = "test_points",
                   data = test_data,
                   overwrite = TRUE)
#> Duckdb spatial extension installed and loaded
#> # Source:   table<test_points> [10 x 4]
#> # Database: DuckDB v0.9.2 [unknown@Linux 6.2.0-1018-azure:R 4.3.2/:memory:]
#>        x     y    id geom      
#>    <int> <int> <int> <list>    
#>  1     1     1     1 <raw [32]>
#>  2     2     2     2 <raw [32]>
#>  3     3     3     3 <raw [32]>
#>  4     4     4     4 <raw [32]>
#>  5     5     5     5 <raw [32]>
#>  6     6     6     6 <raw [32]>
#>  7     7     7     7 <raw [32]>
#>  8     8     8     8 <raw [32]>
#>  9     9     9     9 <raw [32]>
#> 10    10    10    10 <raw [32]>

# file
createDbSpatPoints(conn = duckdb_conn,
                   tbl = "test_points",
                   data = 'test_data.csv',
                   overwrite = TRUE)
#> Duckdb spatial extension installed and loaded
#> # Source:   table<test_points> [10 x 4]
#> # Database: DuckDB v0.9.2 [unknown@Linux 6.2.0-1018-azure:R 4.3.2/:memory:]
#>        x     y    id geom      
#>    <dbl> <dbl> <dbl> <list>    
#>  1     1     1     1 <raw [32]>
#>  2     2     2     2 <raw [32]>
#>  3     3     3     3 <raw [32]>
#>  4     4     4     4 <raw [32]>
#>  5     5     5     5 <raw [32]>
#>  6     6     6     6 <raw [32]>
#>  7     7     7     7 <raw [32]>
#>  8     8     8     8 <raw [32]>
#>  9     9     9     9 <raw [32]>
#> 10    10    10    10 <raw [32]>