Determine if two dbSpatial objects are disjoint

https://postgis.net/docs/ST_Disjoint.html

Usage

st_disjoint(
  g1,
  g1_geom_colName = "geom",
  g1_cols_keep = "all",
  g2,
  g2_geom_colName = "geom",
  g2_cols_keep = "all",
  overwrite = FALSE,
  output_tblName = "disjoint_geom"
)

Arguments

g1

A dbSpatial object.

g1_geom_colName

character. The geometry column name in g1. default: "geom"

g1_cols_keep

character vector. The column names in g1 to keep after the operation. default: 'all' for all columns

g2

dbSpatial object.

g2_geom_colName

character. The geometry column name in g2. default: "geom"

g2_cols_keep

character vector. The column names in g2 to keep after the operation. default: 'all' for all columns

overwrite

logical. If TRUE, overwrite an existing table with the same output_tblName. default: FALSE

output_tblName

character. The name of the table to store the resulting data. default: "disjoint_geom"

Value

tbl_dbi

See also

Examples

con = DBI::dbConnect(duckdb::duckdb(), ":memory:")

coordinates <- data.frame(x = c(100, 200, 300), y = c(500, 600, 700))
attributes <- data.frame(id = 1:3, name = c("A1", "B1", "C1"))

# Combine the coordinates and attributes
dummy_data <- cbind(coordinates, attributes)
 
points <- dbSpatial(conn = con,
                    name = "points", 
                    value = dummy_data, 
                    overwrite = TRUE, 
                    x_colName = "x", 
                    y_colName = "y")

# preview                     
points |> 
  dplyr::mutate(geom_text = ST_AsText(geom))
#> # Source:   SQL [3 x 6]
#> # Database: DuckDB v0.10.0 [unknown@Linux 6.5.0-1016-azure:R 4.3.3/:memory:]
#>       x     y    id name  geom       geom_text      
#>   <dbl> <dbl> <int> <chr> <list>     <chr>          
#> 1   100   500     1 A1    <raw [32]> POINT (100 500)
#> 2   200   600     2 B1    <raw [32]> POINT (200 600)
#> 3   300   700     3 C1    <raw [32]> POINT (300 700)
                    
# Create a second set of points, with B1 and C1 translated by + 100
dummy_data2 <- dummy_data
dummy_data2[c(2,3),c(1,2)] <- dummy_data2[c(2,3),c(1,2)] + 150
dummy_data2$name <- c('A2', 'B2', 'C2')

points2 <- dbSpatial(conn = con,
                    name = "points2", 
                    value = dummy_data2, 
                    overwrite = TRUE, 
                    x_colName = "x", 
                    y_colName = "y")
# preview                     
points2 |> 
  dplyr::mutate(geom_text = ST_AsText(geom))
#> # Source:   SQL [3 x 6]
#> # Database: DuckDB v0.10.0 [unknown@Linux 6.5.0-1016-azure:R 4.3.3/:memory:]
#>       x     y    id name  geom       geom_text      
#>   <dbl> <dbl> <int> <chr> <list>     <chr>          
#> 1   100   500     1 A2    <raw [32]> POINT (100 500)
#> 2   350   750     2 B2    <raw [32]> POINT (350 750)
#> 3   450   850     3 C2    <raw [32]> POINT (450 850)

res <- st_disjoint(g1 = points, 
                   g1_cols_keep = c("name"), 
                   g2 = points2,
                   overwrite = TRUE)

res
#> # Source:   table<disjoint_geom> [8 x 6]
#> # Database: DuckDB v0.10.0 [unknown@Linux 6.5.0-1016-azure:R 4.3.3/:memory:]
#>   name      x     y    id name_1 geom      
#>   <chr> <dbl> <dbl> <int> <chr>  <list>    
#> 1 A1      350   750     2 B2     <raw [32]>
#> 2 A1      450   850     3 C2     <raw [32]>
#> 3 B1      100   500     1 A2     <raw [32]>
#> 4 B1      350   750     2 B2     <raw [32]>
#> 5 B1      450   850     3 C2     <raw [32]>
#> 6 C1      100   500     1 A2     <raw [32]>
#> 7 C1      350   750     2 B2     <raw [32]>
#> 8 C1      450   850     3 C2     <raw [32]>