Translate x, y coordinates by delta x, delta y for point geometries

Translate x, y coordinates by delta x, delta y for point geometries

Usage

st_translate(tbl, geomName = "geom", dx, dy)

Arguments

tbl

name of a table in a duckdb database

geomName

name of the column containing the geometry value in the tbl. default = "geom".

dx

delta x

dy

delta y

Value

a duckdb table with translated geometries

See also

Examples

con = DBI::dbConnect(duckdb::duckdb(), ":memory:")

coordinates <- data.frame(x = c(100, 200, 300), y = c(500, 600, 700))
attributes <- data.frame(id = 1:3, name = c("A", "B", "C"))

# Combine the coordinates and attributes
dummy_data <- cbind(coordinates, attributes)
 
points <- dbSpatial(conn = con,
                    name = "points", 
                    value = dummy_data, 
                    overwrite = TRUE, 
                    x_colName = "x", 
                    y_colName = "y")

points |>
 dplyr::mutate(geom_text = ST_AsText(geom))
#> # Source:   SQL [3 x 6]
#> # Database: DuckDB v1.0.0 [unknown@Linux 6.5.0-1023-azure:R 4.4.1/:memory:]
#>       x     y    id name  geom       geom_text      
#>   <dbl> <dbl> <int> <chr> <list>     <chr>          
#> 1   100   500     1 A     <raw [32]> POINT (100 500)
#> 2   200   600     2 B     <raw [32]> POINT (200 600)
#> 3   300   700     3 C     <raw [32]> POINT (300 700)

points_translated <- st_translate(tbl = points, dx = 100, dy = -20)

points_translated |>
  dplyr::mutate(geom_text = ST_AsText(geom))
#> # Source:   SQL [3 x 6]
#> # Database: DuckDB v1.0.0 [unknown@Linux 6.5.0-1023-azure:R 4.4.1/:memory:]
#>       x     y    id name  geom       geom_text      
#>   <dbl> <dbl> <int> <chr> <list>     <chr>          
#> 1   100   500     1 A     <raw [32]> POINT (200 480)
#> 2   200   600     2 B     <raw [32]> POINT (300 580)
#> 3   300   700     3 C     <raw [32]> POINT (400 680)