Tessellate a dbSpatial object

Creates a tessellation on the extent of dbSpatial with specified parameters. The tessellation can be composed of hexagons or squares.

Usage

tessellate(
  dbSpatial,
  geomName = "geom",
  name = "tessellation",
  shape = c("hexagon", "square"),
  shape_size = NULL,
  gap = 0,
  radius = NULL,
  overwrite = FALSE,
  ...
)

Arguments

dbSpatial

A dbSpatial object representing the spatial table.

geomName

A character string specifying the name of the geometry column in the database.

name

A character string specifying the name of the tessellation.

shape

A character string indicating the shape of the tessellation. Options are "hexagon" or "square".

shape_size

A numeric value specifying the size of the shapes in the tessellation. If NULL, a default size is calculated. See GiottoClass::tessellate for details.

gap

A numeric value indicating the gap between tessellation shapes. Defaults to 0.

radius

A numeric value specifying the radius for hexagonal tessellation. This parameter is ignored for square tessellations.

overwrite

A logical value indicating whether to overwrite an existing tessellation with the same name. Default: FALSE.

...

Additional arguments passed to underlying tessellation functions.

Value

The function writes the tessellation to the database connected to dbSpatial and returns invisibly.

See also

Examples

coordinates <- data.frame(x = c(100, 200, 300), y = c(500, 600, 700))
attributes <- data.frame(id = 1:3, name = c("A", "B", "C"))

# Combine the coordinates and attributes
dummy_data <- cbind(coordinates, attributes)

# Create a duckdb connection
con = DBI::dbConnect(duckdb::duckdb(), ":memory:")

# Create a duckdb table with spatial points
db_points = dbSpatial(conn = con,
                      value = dummy_data,
                      x_colName = "x",
                      y_colName = "y",
                      name = "foo",
                      overwrite = TRUE)
                      
tessellate(db_points, name = "my_tessellation", shape = "hexagon", shape_size = 60)
#> Error in loadNamespace(x): there is no package called ‘GiottoClass’